requirements.txt
flask
requests
pyngrok
google-generativeai
gunicorn
wfastcgi
setuptools
Pillow
